🐛fix:BOJ_2805_나무자르기 테스트케이스 실패 해결
getTakenLength메소드 내 나무길이를 계속 더하는 result 변수가 범위를 초과하는 테스트 생겨 int에서 long으로 변경

+++ jaea-kim/Week_03/BOJ_2805_나무자르기.java
@@ -12,8 +12,8 @@ public class BOJ_2805_나무자르기 {
      총 가져가는 나무 높이 >= M
      절단기의 최대높이 -> 가져가는 나무 최대한 작게
      */
-    public static int getTakenLength(int mid, int[] treeHeight) {
-        int result = 0;
+    public static long getTakenLength(int mid, int[] treeHeight) {
+        long result = 0;
         for (int i = 0; i < treeHeight.length; i++) {
             if (treeHeight[i] > mid) {
                 result += (treeHeight[i] - mid);
